# API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: SIAR Platform ERP/OMS/WMS API
  description: Enterprise-grade Order Management, Inventory Control, and Fulfillment API
  version: 1.0.0
  contact:
    name: SIAR Digital Support
    email: support@siardigital.com
  license:
    name: Proprietary
    url: https://www.siardigital.com
servers:
  - url: http://localhost:5000/api
    description: Development Server
  - url: https://erp.siardigital.com/api
    description: Production Server
...
paths:
...
  /fulfillment/picking-lists:
    post:
      tags:
        - Fulfillment
      summary: Create picking list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                warehouse_id:
                  type: integer
                order_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: Picking list created
  /fulfillment/packing-lists:
    post:
      tags:
        - Fulfillment
      summary: Create packing list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                picking_list_id:
                  type: integer
                items:
                  type: array
      responses:
        '201':
          description: Packing list created
  /fulfillment/shipments:
    post:
      tags:
        - Fulfillment
      summary: Create shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                packing_list_id:
                  type: integer
                carrier:
                  type: string
                tracking_number:
                  type: string
                estimated_delivery:
                  type: string
                  format: date
      responses:
        '201':
          description: Shipment created

# =====================================================================
# RETURNS & REFUNDS ENDPOINTS (8)
# =====================================================================
...
security:
  - BearerToken: []

# =====================================================================
# COMPONENTS & SCHEMAS
# =====================================================================
...